Update readme to reflect gaen only
Why:
Given that we have split the repos we would like for the project read me
to not have multi-target single repo language as this is no longer the
planned technical strategy.

This commit:
removes references to the GPS build in the readme. A future commit
should include more information about the organization as a whole and
links to to the GPS offering.