Flatten out spec file paths
Why:
Currently we do not have a consistent spec folder structure. Some places
have specs in __test__ folders, others have specs co-located with the
file. We would like to have the folder structures for specs to be
consistent.

Given that we as an organization value and require reliability, we would
like to have good specs covering a majority of the application. And
given that it is easier to tell, at a glance, which files have spec
coverage and which do not when the specs are co-located with the file, we
are choosing to have all specs co-located with their respective
application code.