Remove onboarding is complete from redux persist
Why:
We would like to remove redux from the application as it is currently
used only for holding the onboarding state and is more complexity than
is necessary for the use cases of this app.

This commit:
Replaces the onboarding async storage and state management in redux
persist with a simple context component.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>