Convert entry.js to ts and rename
Why:
Current the Entry.js file is acting as the main navigation component for
the app and has no static analysis coverage. Given the navigation flow
of the app is fundamental for the proper functioning of the app, we
would like for it to be covered by static analysis.

This commit:
Converts the navigators in Entry.js to multiple typescript files.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>