Remove components from tracing strategy
Why:
Previously when we were trying to support both gps and gaen in the same
repo we took a shortcut of using dependency inject of whole components
that were different between the apps, as opposed to refining the data
model such that the ui was agnostic to the tracing strategy data for
sake of time. Now that we are only supporting gaen currently in this
repo we are able to remove the components from the tracing strategy and
use them direct.