Remove content from tracing strategy
Why:
We would like for the tracing strategy to encapsulate only the exposure
data module, that is the BTNativeModule, as this allows us to fully
decouple the data model of the GAEN exposure system and the application
user interface.

This allows us to inject the dependency into
the react app which is a common pattern for this use case and will allow us to easily test
the dependency on the GAEN api and even switch the implementation
completely if necessary.

Additionally, this clear decoupling will allow the two systems to
evolve independently, which we can reasonably expect will be occurring
given the quickly evolving understanding of contact tracing.

This commit:
Removes the application static content from the tracing strategy and
adds the permissionStatusSubscription to it such that we can remove the
direct import of the BTNativeModule from the PermissionsContext.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>