Rename typography to rtl enabled text
Why:
We would like to name the Typography component to be more accurately
named for its intended use, enabling accessable and rtl enabled text.
This will allow us to avoid the name collision with `styles/typography`
which is a more appropriate name for the concept of typographic styles
than a component which handles the view logic of rtl languages.