Remove use prop from rtl enabled text
Why:
We would like to apply styles consistently with the established design
system and without extra overhead of maintaining intermediate
components. That is, we would like to use stylesheets directly.
Currently the RTLEnabledText component has a unnecessary use prop which
we would like to remove as to keep the component as simple as possible