Remove unused util functions
Why:
Currently the utils folder has some extra functions and dependencies
which are not being used.

This commit:
Removes the unused functions and dependencies. We also switched back to
the default linker flag setting for iOS in the xcodeproj as we no longer
need to manually set those for removing gps linked libraries.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>