Fix navigation error when selecting a language
Why:
As a user just downloading the app, I want to be able to select my
preferred language.

This commit:
Fixes a bug where the LanguageSelection screen was removed from the
stack navigator in the OnboardingStack.

This commit also separates out the Screens const into more specific
types for each sub-stack to ensure type safety. Now, all screens
specified in the sub-stack type must be present in order for TypeScript
to compile to avoid this regression in the future.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>