Update storage utility helpers (#15)
Why:
The current getStoreData and setSetStoreData utilities are very generic
so that they can be used for all data types. Since we recently the type
annotations of these functions to return a union type instead of `any`,
we had to do some type gymnastics to make the type signatures align for
more specific use cases, such as storing boolean, string, or null
values.

This commit:
Updates the getStoreData and setStoreData functions to have less
responsibility and not be exported from the utils file in favor of more
specific getters and setters. This has the benefit of being more
explicit about what types are returned from this function, how those
type are encoded, and removes the possibility of storing arbitrary data.
It also consolidates the constants used as async storage keys into one
place, which can serve as a source of truth of what data we are storing
to the device.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>