Broadcast enabled status on entering foreground
Why:
We currently have a bug in which the app will not register if it lost
exposure notification permissions if is backgrounded. So we get into a
state where the app says it has permissions when it does not.

This commit:
Add logic to eamit an event of the current status when the app will be
foregrounded and fixes a bug in the event data structure.

Co-Authored-By: johnschoeman johnschoeman1617@gmail.com