Refactor nav bar on settings screens (#20)
Why: prior to this commit, the screens within the "settings" stack were
using a navigation bar wrapper component instead of the header built in
to React Navigation.

Note: we encountered some issues with getting the screen transitions to
perform correctly when transitioning to or from a screen with a hidden
tab bar. We had to transition back to including modal views within the
top-level navigator. Turns out this is the only way to make a screen up
over the tab bar without buggy transition animations.

See https://github.com/react-navigation/react-navigation/issues/6841 for
more information on this issue.