Fix behavior leaving a string instead of a boolean
Why:
----
When we render the navigators, the logic gate is being run against a
variable that can be assigned as a string if both environment variables
are present. We need to treat it as a boolean check to avoid rendering
errors on the Navigator component.

This Commit:
----
- Casts environment variables check to determine whether to display the
self assessment as boolean