Stop use of component for translations (#54)
Why:
----
We have a script that examines the translation keys used throughout the
application and updates the ones that are new or unused for uploading
and completing the rest of such translations. We would like to use this
script to automate the task of finding unused translation keys but we
are seeing erroneous output when executed.
The reason for it is that the parser is encountering a `Trans` component
that comes directly from the library creators but that the parsers do
not recognize. In our particular case we are seeing that a key like:

```
<0></0><1></1>: <0></0><1></1>
```

Is being added when executing `yarn i18next`. There are [issues]
documenting extraneous behavior with this component. Since the usage we
were giving it was pretty small, we might benefit from not using it and
avoid fighting the issue while gaining confidence in our `yarn i18next`
script.

[issues]: https://github.com/i18next/i18next-parser/issues

This Commit:
----
- Removes the usage of the `Trans` component from i18next
- Re-arrange action button text for EmergencyAssessment screen to
include the styled texts while not making use of the i18next syntax for
the translation component

Reviewers:
----
I worked a bit to make the parser agree with the use of the trans
component but hit a brick wall and couldn't advance further. It appeared
that the use of this component didn't grant all the effort so I opted
for removing it. Happy to continue tinkering further if the benefit
grants it.