Use tracingStrategy to access native module (#45)
Why:
We currently have a TracingStrategyContext that is responsible for
determining whether the tracing strategy is gps or GAEN via the
environment. This is no longer necessary now that this repo is
exclusively uses the GAEN strategy.

This commit:
- Removes the unnecessary TracingStrategyContext
- Updates strategy dependent contexts (like Exposure and Permissions
contexts) to import their strategy directly instead of relying on props.
- Updates direct references to the `nativeModule` to be called via the
tracingStrategy instead.