Add submit debug form screen
Why:
We would like for our users and testers to be able to submit their
exposure history debug logs.

This commit:
Adds a simple form with optional inputs for email and a description with
the exposure notifications diagnostics logs which when submitted will
post to an web api in which we will store the debug info.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>
Co-Authored-By: Matt Buckley <mattthousands@nicethings.com>
Co-Authored-By: Ferrumofomega <sam@pathcheck.org>