Only allow code submission when en are enabled
Why:
Currently, if a user has not enabled exposure notifications for the app
it is possible to try to submit the verification code in the affected
user flow, which will necessarily error as the exposure notification api
must be accessible in order to submit exposure keys.

We would like to make it such that a user cannot submit a verification
code unless the en api is enabled. Additionally, if the user starts the
affected user flow without enabled exposure notifications, they should
be prompted to enable exposure notifications.

This commit:
Adds a conditional to the CodeInputScreen to conditionally render either
the CodeInputForm or an EnableExposureNotifications screen that has a
link to the settings for en notifications.

Co-Authored-By johnschoeman <johnschoeman1617@gmail.com>