Use time range for exposure history timestamp
Why:
----
The actual data for an exposure event is not exact on the moment it
happened. Showing an exact date, or relative date, is not correct and
leads to confusion. The reality is that when we get an exposure history
event, this could have happened anytime between 3 days from the
occurrence reported.

How:
---
We opted for generating 3 buckets to classify the time range for an
exposure event:

- within 3 days
- between 4 and 6 days
- between 7 and 14 days

Any event over 14 days of age will not be reported.

This PR:
----
- Introduces the exposureWindowBucket function to determine the range
where the exposure event falls in.
- Add translations for expressing the window bucket in words

Reviewers:
----
We needed to add a lot of type casting now that the exposure events that
will be shown on the list are exclusively `Possible` events. This will
be addressed in future iterations.

Co-Authored-By: Alejandro Dustet <alejandro@thoughtbot.com>