Update pull and push scripts for lokalise (#75)
Why:
----
The push and pull actions for lokalise are not working as expected

This Commit:
----
- Remove location permission strings for iOS
- Extract to `src` not `app`
- Do not search under the `gps` flavor of android to upload files
- Update README portion of scripts with new env variables names
- Change some translations that mention location
- Add option to pull script to include just a set of languages
- Removed unused strings on iOS
- Upload translations from the src folder
- Ran push and updated master translations file on lokalise

Caveat:
----
This Commit does not include the result of running and commiting the
`pull` script since it brings several files into the `iOS` and `android`
projects that might not be necessary. There are more arguments like the
languages to include in the pull we might add to the script.