Check en permission on app state change
Why:
Currently if the user updates the permission for the app while the app
is backgrounded the application will not receive the onStatusUpdated
event from the native layer, which leads to a situation where the app
shows the incorrect permission status.

This commit:
Adds a subscription to the PermissionsContext using react-natives
AppState api and will check the permission status on every change of the
app state.

Note that we are still finding cases where the checkENPermission()
method is returning ["Authorized", "Enabled"] when we would expect
["Authorized", "Disabled"], which seems to be another bug unrelated to
the backgrounding behavior.

Co-Authored-By: mattthousand <matt@nicethings.io>
Co-Authored-By: amandabeiner <aebeiner@gmail.com>