Update permissions en status and request on home
Why:
There are some inconsistencies in how we are receiving the devices en
permission status.

This commit:
Changes the check of the devices exposureNotificationStatus to check
against the status directly vs. via the exposureNotificationsEnabled
method. We also directly request permissions in the app when the app is
unauthorized rather than making the user open the os settings and
manually updating the permissions.

Co-Authored-By: mattthousand <matt@nicethings.io>