Add spec coverage for exposure history flow
Why:
We recently changed the exposure history flow from a calendar interface
to a list of possible exposures. We want to ensure we have spec coverage
so that we can be confident in these changes.

This commit:
- Renames ExposureHistory/History/History.tsx and
ExposureHistory/History/History.spec.tsx to be `index.tsx` files
- Adds specs for the ExposureDetail and NoExposures screens
- Adds specs to the ExposureDetail screen to ensure that the date bucket
of the exposure is accurate