Accept configuration for available locales
Why:
----
The language selection screen has a static group of available locales.
Those are the same locales available for the users of the applications.
This selection needs to be dynamic for each one of the builds of the
applications, since different geographic zones will have specific
translation needs.

How:
----
We need to add an environment variable that can be set per HA to control
which translations will be made available for each. With this change the
initialization of the translation service can not be made by default. It
needs to happen as the application starts.

This Commit:
----
- Initialize i18Next on demand when the application starts not when the
`languages` file loads
- Set the available locales from a configuration variable
- Fall-back to `en` if the provided locales are not valid or not
available

Reviewers:
----
The way we are testing the application by checking the real translation
value forces us to initialize the translation service with the test
runner. This might seem like an overreach for the purpose of testing the
units, but this was the way it was done before and I felt it was out of
the scope of this PR. We can address it later and decide if it is
correct for us to test the translation values vs a mock of the service
that could print the key it receives.

Co-authored-by: johnschoeman <johnschoeman1617@gmail.com>