Add test for hmac key calculation helper
Why:
----
This helper file was not tested and holds a pretty important part of the
logic in sending the data to be verified.

This PR:
----
- Remove unnecessary mock of react-native-simple-crypto
- Transform simple crypto for testing
- Add test for hmac calculation method

Reviewers:
----
The method has quite a few mocks since most of the logic goes through
calling the simple crypto library for encryption and random values
generation purposes.