Add tests for verification API client
Why:
----
The client code for the verification api, that performs code
verification and posts the certificate and hmac digest to the GAEN
server is lacking test coverage. We need to add them so we can add more
checks and refactor with confidence.

This Commit:
----
- Add tests for failure and success response for the verify code
endpoint client method
- Add tests for failure and success response for the certificate
endpoint client method