Move post diagnosis keys to js layer
Why:
----
The logic to submit keys to the backend is duplicated on the native
layer.
We would like to consolidate all the network calls to the js layer, in
order to make all those calls easier to test, instrument and maintain.
We need to display the errors generated on the servers, so it is no
longer a generic error message being displayed.

This Commit:
----
- Add a context to temporarily hold exposure keys for publishing
- Remove the call to the submit diagnosis keys on the native layer
- Add the post diagnosis keys request to the exposureNotificationAPI
module
- Add error handling and complete response logging

Co-authored-by: Alejandro Dustet <alejandro@thoughtbot.com>