Verify that Greek translations are working
Why:
As a product intended for users in Greece and/or Cyprus, we want to make
sure that the Greek translations are working on the exposure history
page, and that our translations are up to date with Lokalise.

This commit:
Ran the i18n update method to make sure we had the latest translations,
and confirmed that the Greek translations are appearing on the
ExposureHistory flow.

Note that we are still missing some translations (they are Empty in
Lokalise) for this part of the app. But we have verified that the
ExposureHistory screen is correctly set up to receive Greek translations
since the headline has been correctly translated.