Upgrade react testing library and jest
Why:
We would like to be able to upgrade the react native version from 61 to
62. This upgrade affected the behavior of the test suite such that we
    also need to upgrade react testing library to be compatible with
react native 62.

This commit:
Upgrades testing-library/react-native from 5 to 7, which included
changing `wait` to `waitFor` and replacing the `{ exact: false }` option
for the `getByText` matcher with a regex. We also updated jest from 25
to 26 as a matter of best practice, which had not breaking changes.

Co-Authored-By: johnschoeman <johnschoeman1617@gmail.com>