Remove unused exposure types (#157)
Why:
Previously we had the need to build an exposure history that had needed
to represent days without exposures. Since we are now using a list view
to represent a users exposure history these data types are no longer
needed.

This commit:
Refactors the union type of an ExposureDatum to be an interface.