Wrap Code Input Form in a SafeAreaView (#166)
Why?
------
Cancel button is hidden when the keyboard opens and the number pad is not dismissible. The user is stuck without a way to cancel out of this screen without a valid code. Instead of adding a SafeAreaView here, we decided it was okay that the cancel is hidden while a user inputs a code, so long as the user also has the ability to dismiss the keyboard and cancel at any point.

This Commit
------
Add an onSubmitEditing that dismisses the number pad when pressing done.