Update Bluetooth status on home screen on toggle
Why:
As a user, who has been prompted to enable bluetooth, I want to see my
new bluetooth status reflected in the home screen so that I know it is
working.

This commit:
Introduces the AppState listener from react native. When an iOS user
foregrounds the app (in order to activate bluetooth in their settings),
or an Android user focuses the app (in order to either activate their
bluetooth settings, or use the notification menu to toggle bluetooth),
the app will now refetch the bluetooth status and set it into state.

This also fixes a bug I experienced where the status was not being
properly set on Android, because `isBluetoothEnabled` returned a
boolean, and we were checking if `status === 'true'`