Short-circuit key uploading process for empty set
Why:
----
If no keys are uploaded to the exposure notifications server, the
process raise an error for an empty set of keys. Once we validate a test
code, we should stop the process from continuing since no data will
cause an error.

This Commit:
----
- Navigate to the affected user flow completed screen if no exposure
keys are available to be uploaded on the device