Remove eula from activation stack
Why:
We would like to suppress the EULA for now until the product
requirements for it can be fully defined.

This commit:
Refactors the activation stack and removes the eula screen. We kept the
eula component in the repo as we expect to be reintroducing it in an
upcoming iteration.