Use the application name consistently
Why:
----
We have an IN_APP_NAME variable that will not be necessary, the app anme
should be uniform.
For more context, the welcome message will become it's own configuration
and the IN_APP_NAME on the privacy part will be removed.
This goes with a [PR] on the resources repo

This Commit:
----
- Removes usage of the IN_APP_NAME environment variable

[PR]: https://github.com/Path-Check/pathcheck-mobile-resources/pull/13