Add Eula and Privacy Policy screen
Why:
As a user, I want to be able to read and agree to my Health Authority's
Privacy Policy and EULA if they have them.

This commit:
Reintroduces the UI to accept the EULA as part of the activation flow,
and updates it according to the new design specs.

Not that we are still not exposing the EULA screen. The UI has been
updated, but we will have to wait to expose it until we implement a
solution for customizing the EULA per jurisdiction