Add bugsnag and logs to the key submission process (#244)
Why:
----
We are lacking visibility on the data flow during the submission of
exposure keys. We need a way to gain insight on what are the errors, on
which environment and what are the problems we are having. Namely which
exchange is happening among all actors in the process:
- OS and application with keys
- Verification server and application
- Exposure Keys server and application

How:
----
By adding a logs service like `bugsnag` we not only have the ability to
log the data flow among all the actors but also gain info on how the
users are behaving and what other errors we might be missing.

This Commit:
----
- Install and setup `bugsnag`
- Add `Logger` service object to add relevant metadata to the session
and send errors
- Add logs to the submission process and fix tests

Co-authored-by: Alejandro Dustet <alejandro@thoughtbot.com>

Co-authored-by: John Schoeman <johnschoeman1617@gmail.com>