Add scaffolding for Analytics work (#242)
Why:
As a product team, I want to be sure that the analytics information I am
collecting respects user privacy and jurisdictional restrictions.

This commit:
Introduces an AnalyticsContext which will will access information
about a user's analytics settings. Right now this is just whether they
have consented to send us analytics, but in the future this will include
more granular privacy settings (ability to consent to some analytics but
not others), and blanket restrictions for jurisdictions (such as HAs in
the EU opting out of analytics altogether)

It also introduces an analytics module, which will serve as a source of
truth for what kinds of events we support, and how to communicate with
the server to fire off those events. This is pretty empty right now, and
is currently serving only to test that we are only tracking analytics
when we have permission to do so.

Note that this code is currently not in use anywhere. We will build off
of it once we finalize tracking requirements with the product and server
teams.