Add Script to get icon assets resources repo
Why:
----
We need to update the icons for each one of the builds corresponding to
a different jurisdiction.

Caveat:
----
This process has only been tested locally and is still to test on the CI
environment. Also there is currently only on jurisdiction with the
assets available.

This Commit:
----
- Add a script to download assets from resources repo and copy to
locations inside ios/android projects

Co-authored-by: John Schoeman <johnschoeman1617@gmail.com>