Add configuration context (#239)
Why:
----
The configuration logic needs several safeguards on the way it is being
used across the application, from defensive programming to decisions on
how to consume it. Some of the operations are changing rapidly and it is
hard to keep up. Also, the `react-native-config` package is very hard to
test properly. To sum up:

We need a way to abstract the way we get configuration items for the
application. This will be an item subject to changes in the next
iterations. A few problems we are facing:

- The complexity of "interpreting" environment is scattered across the
codebase
- Some logic is leaking from the environment into the business logic
- If any environment variable triggered a change in behavior testing
that becomes cumbersome
- Lack of types from environment variables, everything is a string

How:
----
This is attempting moving that logic from just reading string variables
to moving the repeated logic decisions into a domain we can control.
Right now is just decisions on what platform dictates what environment
variable to load, but if this is a pattern we'll keep using is best to
move it to a "testable" and more controlled area. We also anticipate
that more "decisions" will be made in the load phase of a configuration.

Caveats:
----
- The `helpers` are not included in this first pass since it would
involve passing arguments since these are not components. Still working
through the best way to send these configs only used in helpers.
Comments welcome!
- We are still missing a validation of a config for a given HA during
build time. That is coming in subsequent PR's as we work through the
best way of doing this on CI and fastlane.

This Commit:
----
- Migrate Home screen to use the new exposure context
- Add region codes and app package name to context
- Move usage of env on about and legal screen
- Move usage of env on the main tab navigator
- Renamed variables to include `helathAuthority`
- Move usage of env on More screens to use configuration context
- Remove env from ExposureHistory screens to use the context
- Remove usage of env from Share screen on self assessment
- Transform self assessment share to typescript