Use single share app link
Why:
Currently we are making a distinction between ios and android share
links, which doesnt meet the feature requirements as we do not know the
device of the user inwhich the app is being shared.

This commit:
Removes the casing on the device platform for the share app link and
replaces the two links in the env file with a single 'SHARE_APP_LINK'
If the env file does not provide a share app link, we do not render the
share app cta on the home screen.

Co-Authored-By: Arthur Gibson <arthur.gibson@pathcheck.org>