Convert splash screen and download icon asset
Why:
----
The splash screen for the app should be a blank screen with the logo of
the app centered.
The status bar on android has a solid color, that is contrasting with
the color from the top section on the views.

How:
----
The resource for ios that does this is the LaunchScreen, so it needs to
be modified to a white canvas with a centered image. In this case the
loglogoImage asset. For android is the launch_screen resource, that
needs to be transformed in the same way.

The new resources will be available from the resources repo, and placed
in the right spot by the download assets script.

This Commit:
----
- Refactor helpers to download files
- Convert ios splash screen to be a centered logo
- Update android splash screen to be a white background with the
centered logo
- Copy ios logoImage to project from resources repo
- Use fixed colors for main layout on android
- Add splash background color
- Make status bar transparent for android
- Make the color for status bar transparent
- Delete unused bgImage

Co-authored-by: Thomas Paresi <thomas.paresi@thoughtbot.com>