Run download assets with the set_ha script
Why:
----
We need an automated way on CI to setup the right assets for the
different jurisdictions. Currently the set_ha script is ran on each
release build on Bitrise.

This Commit:
----
- Add `download_assets` to the set_ha script

Caveat:
----
If the script is executed for one of the jurisdictions that have all the
assets ready, it will create a huge number of extra files that will
pollute the local environment. We could change the script settings for
local setup purposes(adding an env var to the .env file maybe), but for
practical purposes we do need the assets to be copied over on the CI
environment

Co-authored-by: Thomas Paresi <thomas.paresi@thoughtbot.com>