Add helper for displaying configurable copy per HA
Why:
----
There are three configurable options on the application that each
authority will set up independently, those user-facing text bodies need
to be localized and should follow the same language preference that the
application is using. They need to live on the resources repo since they
are private to each HA.

This Commit:
----
- Add the download copy script
- Add the helper to consume a key, and fetch it from the copy json
- Add default values to useAuthorityCopy
- Display content from authority copy on legal and about if present
- Change surrounding scrollview for legal and about screens
- Update download file script
- Include download_copy_methods on the set_ha script

Reviewers:
----
This approach assumes that the content of the copy file will be verified
prior usage. Also assumes that the supported locales are in sync with
the ones that are present on the specific authority configuration.
The approach follows constraints expressed on the trello card. The new
version respects the current behavior and should not change anything
except for the content of the screen after the app is built per
authority.

Co-authored-by: Thomas Paresi<thomas.paresi@thoughtbot.com>
Co-authored-by: johnschoeman <johnschoeman1617@gmail.com>