Update response handling for keys posting (#300)
* Update response handling for keys posting

Why:
----
When the server returns a response stating that the keys sent were
already inserted but no revision token was sent, we are displaying a
scary and hard to understand error. This operation should not be treated
as an error since the keys were added and therefore the exposed devices
were notified and it should happen only if a positive was reported with
a different jurisdiction. There is a disproportion of occurrences of it
given the nature of how we test.

This Commit:
----
- Modify request management to handle no-op responses
- Add some copy to the error for the no-op response to the users
- Add kind attribute to avoid casting