Display authority links on about and legal screen
Why:
----

Authorities will need to display a set of links under the
about and legal screens information.

How:
----

By using the values on the new overrides file
`configuration/links.json`, we will go through the contents of each file
and display the links under the required screen.

This Commit:
----
- Add useAuthorityLinks helper
- Add tests and a default locale to the useAuthorityCopy hook
- Display authority links on legal and about screens
- Add download links methods for CI

Caveat:
----

This depends on the PR's for adding these necessary configuration files