Add scrollview around welcome screen
Why:
----
On smaller phones the call to action on the welcome screen can not be
tapped because the image and text push it out the screen. We need this
button to be reachable by users when it is not reachable.

How:
----
Adding a scroll view container where the container can grow to its full
height will allow users on smaller screens to scroll down as they need
to.

This Commit:

- Add scroll view container to welcome screen