Validate exposure keys from native module
Why:
Currently we have a bug in our android builds that was caused by the
gaen native module getExposureKeys() function returning empty strings,
this lead to an error in key submission, but we would like to be able to
raise this error and log it sooner in the process.

This commit:
Adds a validate step to NativeModule.getExposureKeys and logs an error
with meta data if the validation fails.

Co-Authored-By: alejandro dustet <alejandro@thoughtbot.com>