Fix bug on fix proximity tracing event (#434)
Why:
Currently there is a bug in iOS 13.7 when a user tries to tap the 'Fix
Proximity Tracing' link on the home screen, where nothing will happen.
This was caused by the NatvieModule.requestAuthorization function
throwing and not being handled.

We would like to try to update the authorization immediately via the
gaen api, and if this fails, we would like to indicate to the user that
the need to update permissions manually via their settings app and link
them to the settings app.

This commit:
Wraps the call to request authorization in a try catch in the component
and shows the Alert dialog with settings link if the request throws.
Additionally we added a logger to the native module function to capture
any failure events in the future. We did not fix the underlying issue of
the `manager.setExposureNotificationEnabled` throwing in the iOS layer.