Improve analytics opt-in behavior
Why:
As a user visiting the settings screen, I want to be able to learn more
about sharing anonymized data.

Currently, if a user has _not_ yet opted into data sharing, clicking
the toggle redirects them to a new screen to turn it on. This is an
unexpected mobile UX pattern, since I would expect toggling a switch to
change the state immediately. When a user _has_ opted into analytics,
the toggle behaves as expected.

This commit:
Updates the Share Anonymized Data List Item to use the same UX pattern
we are currently using on the Home screen. We show the user's current
state with an X or checkmark icon. Clicking on the line item gives them
information about how we use their anonymized data and allows them to
change their status.

Co-Authored-By: johnschoeman <johnschoeman1617@gmail.com>