Remove welcome image downloading script
Why:
----

We need to consolidate the number of scripts we run prior to building
the app. The separation of the welcome image from the rest of in app
image assets is not necessary anymore. The naming on some of the
configurable images should also be reflecting the usage so when we
create the assets the naming is an indication on how will the images be
used

This Commit:
----
- Remove the download_welcome_image script and helpers since the welcome
image will from now on be included on the configurable_images zip file
- Add diagnosis to the image download script to detect miss-matches
during the build process.
- Rename protect privacy images
- Rename info screen images

Reviewers:
----

This depends on a PR on the configuration repo, the resource commit is
updated accordingly for when it is merged.