Standardize naming of Screen types
Why:
----
Prior to this commit, we had some screen types using the pattern
"XScreen" (e.g. "HomeScreen") and others using the pattern
"XStackScreen" (e.g. "HomeStackScreen"). We would like the pattern we
use to be consistent.

This commit:
----
Updates all Screen types to use the pattern "XStackScreen" as in
"HomeStackScreen".