Add SymptomGroup to context
Why:
As a user, I want to know how at risk I am so that I can get accurate
recommendations for my demographic information and symptoms.

This commit:
Adds a symptomGroup to the SelfScreenerContext and a list of possible
symptom groups. As a user fills out the self screener, we use the
useSymptomGroup hook to update the context with their risk category so
that we can show them useful information.

This required a significant change to our GeneralSymptoms and how they
are stored in the context. Though GeneralSymptoms are displayed together
in the UI, they are actually divided into PrimarySymtoms,
SecondarySymptoms and OtherSymptoms, which represent how relevant and
severe they are to determining whether the user at an elevated risk to
COVID-19.

Since the distinction between primary, secondary, and other symptoms is
core to the purpose of the SelfScreener (giving the user personalized
information), and displaying them as one screen is an implementation
detail for the view layer, I opted to change the data structure of the
SelfScreenerContext to separate symptoms by category instead of showing
them all together.

Though this means that each category of symptom now has its own hook and
setter function, it clarifies the complex conditionals for assigning
symptom groups in the core logic a lot.