Add first draft of covid data dashboard
Why:
----

We believe that insight on the data for the virus is a valuable asset
for our users.

This Commit:
----

- Introduces a covid data component that will pull the data from a
specific API and display the results on the home screen when the module
for it is active from the configuration. It needs a valid state
abbreviation to get the correct data for a given jurisdiction.

Co-authored-by: John Schoeman <johnschoeman1617@gmail.com>