Fix navigation to HowItWorks from Settings
Why: prior to this commit, we were using `navigation.navigate` to go to
the `HowItWorks` stack from `Settings`. This caused the screens to
navigate "back" to the `HowItWorks` stack, which felt strange from a UX
perspective. We would like to navigate "forward" to the `HowItWorks`
stack and then "back" to `Settings` afterwards.

This commit:
- Uses `navigation.push` instead of `navigation.navigate` to go to the
  `HowItWorks` stack from `Settings`. The types for React Navigation do
not incude `push` for some reason, so when we use the `useNavigation`
hook we need to provide the generic argument `any`. There may be a
better way to handle this case.
- Uses `navigation.goBack()` to navigate back to `Settings` after the
  `HowItWorks`stack