Exposure Detection Success/Failure Message (#561)
#### Why:
Currently, when a user taps the `Check for exposures` button on the exposure history screen it updates the `Last updated at` info only.  This is quite subtle and its not always clear to users that the check for exposures happened successfully.  We would like to find a way to indicate to the user whether the check was successful or not, even if there where no new exposures found.

#### This commit:
This commit implements a toast/flash message indicating to the user whether the most recent exposure check was successful or not

![Oct-15-2020 10-11-37](https://user-images.githubusercontent.com/2637355/96141673-3d09be80-0ecf-11eb-8d2c-5d7c1e5af1cf.gif)